backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
alembic==1.14.1
python-multipart==0.0.20
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
